+++ rpcs3/Emu/FS/vfsDevice.cpp
@@ -26,7 +26,7 @@ void vfsDevice::SetPath(const std::string& ps3_path, const std::string& local_pa
 
 u32 vfsDevice::CmpPs3Path(const std::string& ps3_path)
 {
-	const u32 lim = std::min(m_ps3_path.length(), ps3_path.length());
+	const u32 lim = (u32)std::min(m_ps3_path.length(), ps3_path.length());
 	u32 ret = 0;
 
 	for(u32 i=0; i<lim; ++i, ++ret)
@@ -58,10 +58,10 @@ u32 vfsDevice::CmpLocalPath(const std::string& local_path)
 	std::vector<std::string> arr0 = fmt::rSplit(path0.GetFullPath(), DL);
 	std::vector<std::string> arr1 = fmt::rSplit(local_path, DL);
 
-	const u32 lim = std::min(arr0.size(), arr1.size());
+	const u32 lim = (u32)std::min(arr0.size(), arr1.size());
 	u32 ret = 0;
 
-	for(u32 i=0; i<lim; ret += arr0[i++].size() + 1)
+	for(u32 i=0; i<lim; ret += (u32)arr0[i++].size() + 1)
 	{
 		if(fmt::CmpNoCase(arr0[i],arr1[i]) != 0)
 		{
@@ -75,7 +75,7 @@ u32 vfsDevice::CmpLocalPath(const std::string& local_path)
 std::string vfsDevice::ErasePath(const std::string& path, u32 start_dir_count, u32 end_dir_count)
 {
 	u32 from = 0;
-	u32 to = path.length() - 1;
+	u32 to = (u32)path.length() - 1;
 
 	for(uint i = 0, dir = 0; i < path.length(); ++i)
 	{
@@ -89,7 +89,7 @@ std::string vfsDevice::ErasePath(const std::string& path, u32 start_dir_count, u
 		}
 	}
 
-	for(int i = path.length() - 1, dir = 0; i >= 0; --i)
+	for(int i = (int)path.length() - 1, dir = 0; i >= 0; --i)
 	{
 		if(path[i] == '\\' || path[i] == '/' || i == 0)
 		{
@@ -109,9 +109,9 @@ std::string vfsDevice::GetRoot(const std::string& path)
 	//return fmt::ToUTF8(wxFileName(fmt::FromUTF8(path), wxPATH_UNIX).GetPath());
 	if(path.empty()) return "";
 
-	u32 first_dir = path.length() - 1;
+	u32 first_dir = (u32)path.length() - 1;
 
-	for(int i = path.length() - 1, dir = 0, li = path.length() - 1; i >= 0 && dir < 2; --i)
+	for(int i = (int)path.length() - 1, dir = 0, li = (int)path.length() - 1; i >= 0 && dir < 2; --i)
 	{
 		if(path[i] == '\\' || path[i] == '/' || i == 0)
 		{
@@ -139,9 +139,9 @@ std::string vfsDevice::GetRootPs3(const std::string& path)
 
 	static const std::string home = "/dev_hdd0/game/";
 	u32 last_dir = 0;
-	u32 first_dir = path.length() - 1;
+	u32 first_dir = (u32)path.length() - 1;
 
-	for(int i = path.length() - 1, dir = 0; i >= 0; --i)
+	for(int i = (int)path.length() - 1, dir = 0; i >= 0; --i)
 	{
 		if(path[i] == '\\' || path[i] == '/' || i == 0)
 		{
